Update signin notification logic
