+++ src/ui/components/App.js
@@ -1,18 +1,32 @@
 import React, { useState, useEffect } from "react";
 import { connect } from "react-redux";
+import { useApolloClient, ApolloProvider } from "@apollo/client";
+import { gql, useMutation } from "@apollo/client";
+import { useAuth0 } from "@auth0/auth0-react";
+
 import DevTools from "./DevTools";
 import Account from "./Account";
 import Loader from "./shared/Loader";
 import { AppErrors, PopupBlockedError } from "./shared/Error";
 import SharingModal from "./shared/SharingModal";
 import { isDeployPreview } from "ui/utils/environment";
+import { setUserInBrowserPrefs } from "ui/utils/browser";
 import { selectors } from "ui/reducers";
-import { useApolloClient, ApolloProvider } from "@apollo/client";
-import { useAuth0 } from "@auth0/auth0-react";
+import { actions } from "ui/actions";
 import { hasLoadingParam } from "ui/utils/environment";
 import ResizeObserverPolyfill from "resize-observer-polyfill";
-import { actions } from "../actions";
+
 import "styles.css";
+
+const CREATE_SESSION = gql`
+  mutation CreateSession($object: sessions_insert_input!) {
+    insert_sessions_one(object: $object) {
+      id
+      controller_id
+      recording_id
+    }
+  }
+`;
 
 function useGetApolloClient() {
   const [apolloClient, setApolloClient] = useState(null);
@@ -52,9 +66,27 @@ function installViewportObserver({ updateNarrowMode }) {
   observer.observe(viewport);
 }
 
-function App({ theme, recordingId, modal, updateNarrowMode }) {
-  const { isLoading } = useAuth0();
+function App({ theme, recordingId, modal, updateNarrowMode, updateUser, sessionId }) {
+  const auth = useAuth0();
   const { apolloClient, consentPopupBlocked } = useGetApolloClient();
+
+  const [CreateSession] = useMutation(CREATE_SESSION);
+
+  useEffect(() => {
+    setUserInBrowserPrefs(auth.user);
+    updateUser(auth.user);
+  }, [auth.user]);
+
+  useEffect(() => {
+    if (auth.user && sessionId) {
+      const object = {
+        id: sessionId,
+        recording_id: recordingId,
+        controller_id: sessionId.split("/")[0],
+      };
+      CreateSession({ variables: { object } });
+    }
+  }, [auth.user, sessionId]);
 
   useEffect(() => {
     document.body.parentElement.className = theme;
@@ -65,7 +97,7 @@ function App({ theme, recordingId, modal, updateNarrowMode }) {
     return <PopupBlockedError />;
   }
 
-  if ((!isDeployPreview() && isLoading) || !apolloClient || hasLoadingParam()) {
+  if ((!isDeployPreview() && auth.isLoading) || !apolloClient || hasLoadingParam()) {
     return <Loader />;
   }
 
@@ -83,8 +115,10 @@ export default connect(
     theme: selectors.getTheme(state),
     recordingId: selectors.getRecordingId(state),
     modal: selectors.getModal(state),
+    sessionId: selectors.getSessionId(state),
   }),
   {
     updateNarrowMode: actions.updateNarrowMode,
+    updateUser: actions.updateUser,
   }
 )(App);
